Split up pethelper.py into 3 seperate modules in zoo package
